squashfs: widen arithmetic to unsigned int in SQUASHFS_ID_BYTES()
sBlk->no_ids is unsigned short, and so explicitly widen to
unsigned int.

Also add a note on overflow limits in read_fs.c

+++ squashfs-tools/squashfs_fs.h
@@ -208,7 +208,7 @@
 					sizeof(long long))
 
 /* uid lookup table defines */
-#define SQUASHFS_ID_BYTES(A)	((A) * sizeof(unsigned int))
+#define SQUASHFS_ID_BYTES(A)	((unsigned int)(A) * sizeof(unsigned int))
 
 #define SQUASHFS_ID_BLOCK(A)		(SQUASHFS_ID_BYTES(A) / \
 						SQUASHFS_METADATA_SIZE)
